pyproject.toml
[tool.poetry]
name = "wise-nutrition"
version = "0.1.0"
description = "AI-powered application for nutrition knowledge and personalized recipe generation"
authors = ["Andre Machon ajbmachon2@gmail.com"]
readme = "README.md"
packages = [{include = "wise_nutrition"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.5.0"
pydantic-ai = "^0.1.3"
openai = "^1.12.0"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.6.0"
httpx = "^0.27.0"
supabase = "^2.3.0"
vecs = "^0.4.0"
python-dotenv = "^1.0.0"
fastapi = "^0.115.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
wise-nutrition = "wise_nutrition.cli:app"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "A", "C4", "SIM", "TID"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
